+++ Pester.psm1
@@ -3,6 +3,14 @@
 # Changeset: $sha$
 
 $Script:PesterRoot = Split-Path -Path $MyInvocation.MyCommand.Path
+if ($PSVersionTable.PSVersion.Major -ge 3)
+{
+    $script:IgnoreErrorPreference = 'Ignore'
+}
+else
+{
+    $script:IgnoreErrorPreference = 'SilentlyContinue'
+}
 
 . "$PesterRoot\PesterCore.ps1"
 
@@ -37,8 +45,14 @@ Optionally, Pester can generate a report of how much code is covered
 by the tests, and information about any commands which were not
 executed.
 
-.PARAMETER Path
-The path where Invoke-Pester begins to search for test files. The default is the current directory. Aliased 'relative_path' for backwards compatibility.
+.PARAMETER Script
+This parameter indicates which test scripts should be run.
+This parameter may be passed simple strings (wildcards are allowed), or hashtables containing Path, Arguments and Parameters keys.
+If hashtables are used, the Parameters key must refer to a hashtable, and the Arguments key must refer to an array; these will be splatted to the test script(s) indicated in the Path key.
+
+Note:  If the path contains any wildcards, or if it refers to a directory, then Pester will search for and execute all test scripts named *.Tests.ps1 in the target path; the search is recursive.  If the path contains no wildcards and refers to a file, Pester will just try to execute that file regardless of its name.
+
+Aliased to 'Path' and 'relative_path' for backwards compatibility.
 
 .PARAMETER TestName
 Informs Invoke-Pester to only run Describe blocks that match these names.  The strings passed to TestName may contain wildcards
@@ -80,9 +94,14 @@ Invoke-Pester
 This will find all *.Tests.ps1 files and run their tests. No exit code will be returned and no log file will be saved.
 
 .Example
-Invoke-Pester ./tests/Utils*
+Invoke-Pester -Script ./tests/Utils*
 
 This will run all tests in files under ./Tests that begin with Utils and alsocontains .Tests.
+
+.Example
+Invoke-Pester -Script @{ Path = './tests/Utils*'; Parameters = @{ NamedParameter = 'Passed By Name' }; Arguments = @('Passed by position') }
+
+Executes the same tests as in Example 1, but will run them with the equivalent of the following command line:  & $testScriptPath -NamedParameter 'Passed By Name' 'Passed by position'
 
 .Example
 Invoke-Pester -TestName "Add Numbers"
@@ -117,13 +136,16 @@ about_pester
     [CmdletBinding(DefaultParameterSetName = 'LegacyOutputXml')]
     param(
         [Parameter(Position=0,Mandatory=0)]
-        [Alias('relative_path')]
-        [string]$Path = ".",
+        [Alias('Path', 'relative_path')]
+        [object[]]$Script = '.',
+
         [Parameter(Position=1,Mandatory=0)]
         [Alias("Name")]
         [string[]]$TestName,
+
         [Parameter(Position=2,Mandatory=0)]
         [switch]$EnableExit,
+
         [Parameter(Position=3,Mandatory=0, ParameterSetName = 'LegacyOutputXml')]
         [string]$OutputXml,
 
@@ -136,6 +158,7 @@ about_pester
         [switch]$PassThru,
 
         [object[]] $CodeCoverage = @(),
+
         [Switch]$Strict,
 
         [Parameter(Mandatory = $true, ParameterSetName = 'NewOutputSet')]
@@ -144,6 +167,7 @@ about_pester
         [Parameter(Mandatory = $true, ParameterSetName = 'NewOutputSet')]
         [ValidateSet('LegacyNUnitXml', 'NUnitXml')]
         [string] $OutputFormat,
+
         [Switch]$Quiet
     )
 
@@ -162,34 +186,37 @@ about_pester
 
     $script:mockTable = @{}
 
-    $pester = New-PesterState -Path (Resolve-Path $Path) -TestNameFilter $TestName -TagFilter ($Tag -split "\s") -ExcludeTagFilter ($ExcludeTag -split "\s") -SessionState $PSCmdlet.SessionState -Strict:$Strict -Quiet:$Quiet
+    $pester = New-PesterState -TestNameFilter $TestName -TagFilter ($Tag -split "\s") -ExcludeTagFilter ($ExcludeTag -split "\s") -SessionState $PSCmdlet.SessionState -Strict:$Strict -Quiet:$Quiet
     Enter-CoverageAnalysis -CodeCoverage $CodeCoverage -PesterState $pester
 
     Write-PesterStart $pester
     
-    # TODO:  Make sure this logic is accounted for in Write-PesterStart for the merge.
+    $invokeTestScript = {
+        param (
+            [Parameter(Position = 0)]
+            [string] $Path,
 
-    # $message = "Executing all tests in '$($pester.Path)'"
-    # if ($TestName) { $message += " matching test name '$TestName'" }
-    # 
-    # Write-Screen $message
+            [object[]] $Arguments = @(),
+            [System.Collections.IDictionary] $Parameters = @{}
+        )
 
-    $scriptBlock = { & $args[0] }
-    Set-ScriptBlockScope -ScriptBlock $scriptBlock -SessionState $PSCmdlet.SessionState
+        & $Path @Parameters @Arguments
+    }
 
-    Get-ChildItem $pester.Path -Filter "*.Tests.ps1" -Recurse |
-    where { -not $_.PSIsContainer } |
-    foreach {
-        $testFile = $_
+    Set-ScriptBlockScope -ScriptBlock $invokeTestScript -SessionState $PSCmdlet.SessionState
 
+    $testScripts = ResolveTestScripts $Script
+
+    foreach ($testScript in $testScripts)
+    {
         try
         {
-            & $scriptBlock $testFile.PSPath
+            & $invokeTestScript -Path $testScript.Path -Arguments $testScript.Arguments -Parameters $testScript.Parameters
         }
         catch
         {
             $firstStackTraceLine = $_.ScriptStackTrace -split '\r?\n' | Select-Object -First 1
-            $pester.AddTestResult("Error occurred in test script '$($testFile.FullName)'", "Failed", $null, $_.Exception.Message, $firstStackTraceLine)
+            $pester.AddTestResult("Error occurred in test script '$($testScript.Path)'", "Failed", $null, $_.Exception.Message, $firstStackTraceLine)
             $pester.TestResult[-1] | Write-PesterResult
         }
     }
@@ -199,15 +226,14 @@ about_pester
     Write-CoverageReport -CoverageReport $coverageReport
     Exit-CoverageAnalysis -PesterState $pester
 
-
-    if($OutputFile) {
+    if(Get-Variable -Name OutputFile -ValueOnly -ErrorAction $script:IgnoreErrorPreference) {
         Export-PesterResults -PesterState $pester -Path $OutputFile -Format $OutputFormat
     }
 
     if ($PassThru) {
         #remove all runtime properties like current* and Scope
         $properties = @(
-            "Path","TagFilter","ExcludeTagFilter","TestNameFilter","TotalCount","PassedCount","FailedCount","SkippedCount","PendingCount","Time","TestResult"
+            "TagFilter","ExcludeTagFilter","TestNameFilter","TotalCount","PassedCount","FailedCount","SkippedCount","PendingCount","Time","TestResult"
 
             if ($CodeCoverage)
             {
@@ -219,6 +245,93 @@ about_pester
     }
 
     if ($EnableExit) { Exit-WithCode -FailedCount $pester.FailedCount }
+}
+
+function ResolveTestScripts
+{
+    param ([object[]] $Path)
+
+    $resolvedScriptInfo = @(
+        foreach ($object in $Path)
+        {
+            if ($object -is [System.Collections.IDictionary])
+            {
+                $unresolvedPath = Get-DictionaryValueFromFirstKeyFound -Dictionary $object -Key 'Path', 'p'
+                $arguments      = @(Get-DictionaryValueFromFirstKeyFound -Dictionary $object -Key 'Arguments', 'args', 'a')
+                $parameters     = Get-DictionaryValueFromFirstKeyFound -Dictionary $object -Key 'Parameters', 'params'
+
+                if ($unresolvedPath -isnot [string] -or $unresolvedPath -notmatch '\S')
+                {
+                    throw 'When passing hashtables to the -Path parameter, the Path key is mandatory, and must contain a single string.'
+                }
+
+                if ($null -ne $parameters -and $parameters -isnot [System.Collections.IDictionary])
+                {
+                    throw 'When passing hashtables to the -Path parameter, the Parameters key (if present) must be assigned an IDictionary object.'
+                }
+            }
+            else
+            {
+                $unresolvedPath = [string] $object
+                $arguments      = @()
+                $parameters     = @{}            
+            }
+
+            if ($unresolvedPath -notmatch '[\*\?\[\]]' -and
+                (Test-Path -LiteralPath $unresolvedPath -PathType Leaf) -and
+                (Get-Item -LiteralPath $unresolvedPath) -is [System.IO.FileInfo])
+            {
+                $extension = [System.IO.Path]::GetExtension($unresolvedPath)
+                if ($extension -ne '.ps1')
+                {
+                    Write-Error "Script path '$unresolvedPath' is not a ps1 file."
+                }
+                else
+                {
+                    New-Object psobject -Property @{
+                        Path       = $unresolvedPath
+                        Arguments  = $arguments
+                        Parameters = $parameters
+                    }
+                }
+            }
+            else
+            {
+                # World's longest pipeline?
+
+                Resolve-Path -Path $unresolvedPath |
+                Where-Object { $_.Provider.Name -eq 'FileSystem' } |
+                Select-Object -ExpandProperty ProviderPath |
+                Get-ChildItem -Include *.Tests.ps1 -Recurse |
+                Where-Object { -not $_.PSIsContainer } |
+                Select-Object -ExpandProperty FullName -Unique |
+                ForEach-Object {
+                    New-Object psobject -Property @{
+                        Path       = $_
+                        Arguments  = $arguments
+                        Parameters = $parameters
+                    }
+                }
+            }
+        }
+    )
+
+    # Here, we have the option of trying to weed out duplicate file paths that also contain identical
+    # Parameters / Arguments.  However, we already make sure that each object in $Path didn't produce
+    # any duplicate file paths, and if the caller happens to pass in a set of parameters that produce
+    # dupes, maybe that's not our problem.  For now, just return what we found.
+
+    $resolvedScriptInfo
+}
+
+function Get-DictionaryValueFromFirstKeyFound
+{
+    param ([System.Collections.IDictionary] $Dictionary, [object[]] $Key)
+
+    foreach ($keyToTry in $Key)
+    {
+        if ($Dictionary.Contains($keyToTry)) { return $Dictionary[$keyToTry] }
+    }
 }
 
 function Set-ScriptBlockScope
@@ -260,22 +373,14 @@ function Get-ScriptBlockScope
     [scriptblock].GetProperty('SessionStateInternal', $flags).GetValue($ScriptBlock, $null)
 }
 
-function Get-IgnoreErrorPreference
+$snippetsDirectoryPath = "$PSScriptRoot\Snippets"
+if ((Test-Path -Path Variable:\psise) -and ($null -ne $psISE) -and ($PSVersionTable.PSVersion.Major -ge 3) -and (Test-Path $snippetsDirectoryPath))
 {
-    if ($PSVersionTable.PSVersion.Major -ge 3)
-    {
-        return 'Ignore'
-    }
-    else
-    {
-        return 'SilentlyContinue'
-    }
-
+    Import-IseSnippet -Path $snippetsDirectoryPath
 }
 
 Export-ModuleMember Describe, Context, It, In, Mock, Assert-VerifiableMocks, Assert-MockCalled
 Export-ModuleMember New-Fixture, Get-TestDriveItem, Should, Invoke-Pester, Setup, InModuleScope, Invoke-Mock
 Export-ModuleMember BeforeEach, AfterEach, BeforeAll, AfterAll
 Export-ModuleMember Get-MockDynamicParameters, Set-DynamicParameterVariables
-Export-ModuleMember Get-IgnoreErrorPreference
 Export-ModuleMember Invoke-Gherkin, When -Alias And, But, Given, Then