+++ build.psake.ps1
@@ -3,24 +3,25 @@ properties {
     $currentDir = resolve-path .
     $Invocation = (Get-Variable MyInvocation -Scope 1).Value
     $baseDir = $psake.build_script_dir
-    $version = git describe --abbrev=0 --tags
+    $version = git.exe describe --abbrev=0 --tags
     $nugetExe = "$baseDir\vendor\tools\nuget"
+    $targetBase = "tools"
 }
 
-Task default -depends Build
-Task Build -depends Test, Package
+Task default -depends Test, Build
+Task Build -depends Package
 Task Package -depends Version-Module, Pack-Nuget, Unversion-Module
 Task Release -depends Build, Push-Nuget
 
 Task Test {
-    CD "$baseDir"
+    Set-Location "$baseDir"
     exec {."$baseDir\bin\Pester.bat"}
-    CD $currentDir
+    Set-Location $currentDir
 }
 
 Task Version-Module{
-    $v = git describe --abbrev=0 --tags
-    $changeset=(git log -1 $($v + '..') --pretty=format:%H)
+    $v = git.exe describe --abbrev=0 --tags
+    $changeset=(git.exe log -1 $($v + '..') --pretty=format:%H)
     (Get-Content "$baseDir\Pester.psm1") `
       | % {$_ -replace "\`$version\`$", "$version" } `
       | % {$_ -replace "\`$sha\`$", "$changeset" } `
@@ -28,12 +29,9 @@ Task Version-Module{
 }
 
 Task Unversion-Module{
-    $v = git describe --abbrev=0 --tags
-    $changeset=(git log -1 $($v + '..') --pretty=format:%H)
-    (Get-Content "$baseDir\Pester.psm1") `
-      | % {$_ -replace "$version", "`$version`$" } `
-      | % {$_ -replace "$changeset", "`$sha`$" } `
-      | Set-Content "$baseDir\Pester.psm1"
+    Set-Location $baseDir
+    git.exe checkout -- $baseDir\Pester.psm1
+    Set-Location $currentDir
 }
 
 Task Pack-Nuget {
@@ -44,7 +42,7 @@ Task Pack-Nuget {
     mkdir "$baseDir\build"
     exec {
       . $nugetExe pack "$baseDir\Pester.nuspec" -OutputDirectory "$baseDir\build" `
-      -NoPackageAnalysis -version $version
+      -NoPackageAnalysis -version $version -Properties targetBase=$targetBase
     }
 }